Move ligand-specific code to separate submodule.
Move functions from pdb.py to mol2.py.
