+++ pdb2pqr/pdb2pqr/ligand/__init__.py
@@ -1,0 +1,4 @@
+"""Ligand support functions
+
+Jens Erik Nielsen, University College Dublin 2004
+"""
